touch up side menu bar; modularize icons
